requirements.txt
# Excel
openpyxl>=3.1.0
msoffcrypto-tool>=5.0.0

# Gmail API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# HTTP API
flask>=3.0.0
gunicorn>=21.0.0

# Decryption (PBKDF2 + AES-GCM)
cryptography>=41.0.0

# Config and HTTP client
python-dotenv>=1.0.0
requests>=2.28.0

# SQLite Cloud
sqlitecloud>=0.0.80
